Stats euros to cents

+++ src/main/scala/com/particeep/api/models/transaction/TransactionSearchStatistics.scala
@@ -5,11 +5,11 @@ import ai.x.play.json.Encoders._
 import com.particeep.api.core.Formatter
 
 case class TransactionSearchStatistics(
-    amount:              Double,
-    bare_owner_amount:   Double,
-    fees:                Double,
-    usufructuary_amount: Double,
-    partner_fees:        Double,
+    amount:              Long,
+    bare_owner_amount:   Long,
+    fees:                Long,
+    usufructuary_amount: Long,
+    partner_fees:        Long,
     num_of_shares:       Long
 )
